+++ packages/components/src/components/post-icon/post-icon.tsx
@@ -1,11 +1,7 @@
-import { Component, Element, Host, h, Prop, Watch, Build } from '@stencil/core';
-import {
-  checkEmptyOrType,
-  checkRequiredAndType,
-  checkEmptyOrOneOf,
-} from '@/utils/property-checkers';
-import { version } from '@root/package.json';
 import { ANIMATION_KEYS, PostIconAnimation } from '@/types/icon-animations';
+import { OneOf, Required, Type } from '@/utils';
+import { version } from '@root/package.json';
+import { Build, Component, Element, h, Host, Prop } from '@stencil/core';
 
 const CDN_URL = `https://unpkg.com/@swisspost/design-system-icons@${version}/public/post-icons/`;
 
@@ -24,22 +20,16 @@ export class PostIcon {
   /**
    * The name of the animation.
    */
-  @Prop({ reflect: true }) readonly animation?: PostIconAnimation;
-
-  @Watch('animation')
-  validateAnimation() {
-    checkEmptyOrOneOf(this, 'animation', ANIMATION_KEYS);
-  }
+  @Prop({ reflect: true })
+  @OneOf(ANIMATION_KEYS)
+  readonly animation?: PostIconAnimation;
 
   /**
    * The base path, where the icons are located (must be a public url).<br/>Leave this field empty to use the default cdn url.
    */
-  @Prop() readonly base?: string;
-
-  @Watch('base')
-  validateBase() {
-    checkEmptyOrType(this, 'base', 'string');
-  }
+  @Prop()
+  @Type('string')
+  readonly base?: string;
 
   /**
    * A full URL to the icon file. When set, this property has the highest priority.
@@ -59,32 +49,24 @@ export class PostIcon {
   /**
    * The name/id of the icon (e.g. 1000, 1001, ...).
    */
-  @Prop({ reflect: true }) readonly name!: string;
-
-  @Watch('name')
-  validateName() {
-    checkRequiredAndType(this, 'name', 'string');
-  }
+  @Prop({ reflect: true })
+  @Required()
+  @Type('string')
+  readonly name!: string;
 
   /**
    * The number of degree for the css rotate transformation.
    */
-  @Prop() readonly rotate?: number;
-
-  @Watch('rotate')
-  validateRotate() {
-    checkEmptyOrType(this, 'rotate', 'number');
-  }
+  @Prop()
+  @Type('number')
+  readonly rotate?: number;
 
   /**
    * The number for the css scale transformation.
    */
-  @Prop() readonly scale?: number;
-
-  @Watch('scale')
-  validateScale() {
-    checkEmptyOrType(this, 'scale', 'number');
-  }
+  @Prop()
+  @Type('number')
+  readonly scale?: number;
 
   /**
    * Construct the icon URL according to the following rules:
@@ -165,22 +147,17 @@ export class PostIcon {
   private getStyles() {
     const url = this.getUrl();
 
-    return Object.fromEntries(Object.entries({
-      '-webkit-mask-image': `url(${url})`,
-      'mask-image': `url('${url}')`,
-      'transform':
-        (this.scale && !Number.isNaN(Number(this.scale)) ? 'scale(' + this.scale + ')' : '') +
-        (this.rotate && !Number.isNaN(Number(this.rotate)) ? ' rotate(' + this.rotate + 'deg)' : ''),
-    })
-      .filter(([_key, value]) => value !== null));
-  }
-
-  componentDidLoad() {
-    this.validateBase();
-    this.validateName();
-    this.validateScale();
-    this.validateRotate();
-    this.validateAnimation();
+    return Object.fromEntries(
+      Object.entries({
+        '-webkit-mask-image': `url(${url})`,
+        'mask-image': `url('${url}')`,
+        'transform':
+          (this.scale && !Number.isNaN(Number(this.scale)) ? 'scale(' + this.scale + ')' : '') +
+          (this.rotate && !Number.isNaN(Number(this.rotate))
+            ? ' rotate(' + this.rotate + 'deg)'
+            : ''),
+      }).filter(([_key, value]) => value !== null),
+    );
   }
 
   render() {