+++ packages/internet-header/src/components/post-internet-header/post-internet-header.tsx
@@ -1,15 +1,17 @@
-import { Component, Event, EventEmitter, h, Host, Listen, Prop, Watch } from '@stencil/core';
-import { getLocalizedConfig, isValidProjectId } from '@/services/config.service';
-import { version } from '@root/package.json';
-import { ActiveRouteProp, Environment } from '@/models/general.model';
-import { dispose, state } from '@/data/store';
 import { Link, LinkProps, MegaDropdown, UserMenu } from '@/components/internal';
-import { LinkConfig } from '@/models/shared.model';
+import { dispose, state } from '@/data/store';
+import { ActiveRouteProp, Environment } from '@/models/general.model';
 import { UserMenuConfig } from '@/models/header.model';
+import { LinkConfig } from '@/models/shared.model';
+import { getLocalizedConfig, isValidProjectId } from '@/services/config.service';
+import { getActiveLink } from '@/services/route.service';
+import { version } from '@root/package.json';
+import { Component, Event, EventEmitter, h, Host, Listen, Prop, Watch } from '@stencil/core';
 import '@swisspost/design-system-components';
 
 @Component({
   tag: 'swisspost-internet-header',
+  styleUrl: 'post-internet-header.scss',
   shadow: false,
 })
 export class PostInternetHeader {
@@ -20,8 +22,8 @@ export class PostInternetHeader {
   @Prop() activeRoute: ActiveRouteProp = 'auto';
 
   @Watch('activeRoute')
-  async handleActiveRouteChange() {
-    await this.updateConfig();
+  handleActiveRouteChange() {
+    this.updateActiveUrl();
   }
 
   /**
@@ -86,6 +88,11 @@ export class PostInternetHeader {
   @Prop({ reflect: true }) readonly textMenu!: string;
 
   /**
+   * Visually hidden label for the login widget trigger button.
+   */
+  @Prop({ reflect: true }) readonly textUserMenuTrigger!: string;
+
+  /**
    * Visually hidden label for the user menu.
    */
   @Prop({ reflect: true }) readonly textUserLinks!: string;
@@ -144,6 +151,12 @@ export class PostInternetHeader {
       environment: this.environment,
       language: this.language,
     });
+
+    this.updateActiveUrl();
+  }
+
+  private updateActiveUrl() {
+    state.activeLink = getActiveLink(this.activeRoute);
   }
 
   private renderNavItem(config: LinkConfig | UserMenuConfig, props: LinkProps = {}): string {
@@ -238,7 +251,26 @@ export class PostInternetHeader {
             </post-language-menu>
           )}
 
-          {globalHeader.login && this.renderNavItem(globalHeader.login, { slot: 'post-login' })}
+          {globalHeader.login && (
+            <post-login-widget
+              slot="post-login"
+              textCurrentUser={this.textCurrentUser}
+              textUserMenu={this.textUserLinks}
+              textUserMenuTrigger={this.textUserMenuTrigger}
+            >
+              <Link slot="login-link" config={globalHeader.login} />
+
+              {globalHeader.userMenuLinks && (
+                <div slot="user-links">
+                  {globalHeader.userMenuLinks.map(link => (
+                    <post-menu-item key={link.url}>
+                      <Link config={link} />
+                    </post-menu-item>
+                  ))}
+                </div>
+              )}
+            </post-login-widget>
+          )}
 
           {localHeader.title && <p slot="title">{localHeader.title}</p>}